run mongodb in docker
